% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDINAtree.R
\name{GDINAtree}
\alias{GDINAtree}
\title{G-DINA tree}
\usage{
GDINAtree(
  dat,
  covariates,
  Q,
  model = "GDINA",
  maxdepth = 3,
  minsize = 100,
  alpha = 0.05,
  bonferroni = TRUE,
  ...
)
}
\arguments{
\item{dat}{A \emph{N} individuals x \emph{J} items (\code{matrix} or \code{data.frame}). Missing values need to be coded as \code{NA}.}

\item{covariates}{A \emph{N} individuals x \emph{M} covariates (\code{matrix} or \code{data.frame}). Missing values need to be coded as \code{NA}.}

\item{Q}{A \emph{J} items x \emph{K} attributes Q-matrix (\code{matrix} or \code{data.frame}).}

\item{model}{CDM to be estimated. The possible options include "GDINA","DINA","DINO","ACDM","LLM", "RRUM", "MSDINA", "BUGDINO", "SISM", and "UDF". See the \code{GDINA} package (Ma & de la Torre, 2020) for more information. Default is \code{"GDINA"}.}

\item{maxdepth}{An \code{integer} indicating the maximum depth of a tree.}

\item{minsize}{An \code{integer} indicating the minimum number of observations in a node.}

\item{alpha}{Nominal significance level.}

\item{bonferroni}{A \code{logical} value indicating whether the Bonferroni correction should be applied.}
}
\value{
\code{GDINAtree} returns an object of class \code{GDINAtree}.
\describe{
\item{\code{tree}}{A \code{modelparty} object containing the main results from the G-DINA tree.}
\item{\code{fit.nodes}}{A \code{list} containing the models fitted in each node (each of them being a \code{GDINA} object).}
\item{\code{specifications}}{Function call specifications (\code{list}).}
}
}
\description{
Implementation of a model-based recursive partitioning algorithm (Zeileis et al., 2008) for the G-DINA model (Nájera et al., in press) to detect subpopulations based on
different item parameters, underlying model, or Q-matrix. The function is based on the \code{mob} function of the package \code{partykit} package (Hothorn & Zeileis, 2015), and
estimates CDMs via the \emph{GDINA} package (MA & de la Torre, 2020).
}
\examples{
library(GDINA)
dat <- as.data.frame(rbind(sim30GDINA$simdat, sim30DINA$simdat))
Q <- sim30GDINA$simQ
set.seed(1713)
dat$y <- rep(1:2, each = nrow(dat)/2)
dat$cov1 <- round(runif(nrow(dat)), 2)
dat$cov2 <- round(runif(nrow(dat)), 2)
dat$cov3 <- sample(1:2, nrow(dat), replace = TRUE)
dat$cov4 <- sample(1:4, nrow(dat), replace = TRUE)
fit <- GDINAtree(dat = dat[,1:30], covariates = dat[,31:35], Q = Q)
plot(fit)
}
\references{
Hothorn, T., & Zeileis, A. (2015). partykit: A Modular Toolkit for Recursive Partytioning in R. \emph{Journal of Machine Learning Research}, \emph{16}, 3905-3909. https://jmlr.org/papers/v16/hothorn15a.html

Nájera, P., Sorrel, M. A., Sterner, P., & Goretzko, D. (2026). Invariance Analysis in Cognitive Diagnosis Models with G-DINA Trees. \emph{Manuscript submitted for publication}

Ma, W., & de la Torre, J. (2020). GDINA: An R package for cognitive diagnosis modeling. \emph{Journal of Statistical Software}, \emph{93}(14). https://doi.org/10.18637/jss.v093.i14

Zeileis, A., Hothorn, T., & Hornik, K. (2008). Model-Based Recursive Partitioning. \emph{Journal of Computational and Graphical Statistics}, \emph{17}(2), 492-514. https://doi.org/10.1198/106186008X319331
}
\author{
{David Goretzko, Goethe University Frankfurt, \cr Philipp Sterner, LMU Munich, \cr Pablo Nájera, Universidad Pontificia Comillas}
}
